+++ fakeredis/commands_mixins/connection_mixin.py
@@ -62,7 +62,27 @@ class ConnectionCommandsMixin:
 
     @command(name="CLIENT INFO", fixed=(), repeat=())
     def client_info_cmd(self) -> bytes:
-        return self.client_info
+        return self.client_info_as_bytes
+
+    @command(name="CLIENT LIST", fixed=(), repeat=(bytes,))
+    def client_list_cmd(self, *args: bytes) -> bytes:
+        sockets = self._server.sockets.copy()
+        i = 0
+        filter_ids = set()
+        while i < len(args):
+            if casematch(args[i], b"TYPE") and i + 1 < len(args):
+                i += 2
+            if casematch(args[i], b"ID") and i + 1 < len(args):
+                i += 1
+                while i < len(args):
+                    filter_ids.add(Int.decode(args[i]))
+                    i += 1
+            else:
+                raise SimpleError(msgs.SYNTAX_ERROR_MSG)
+        if len(filter_ids) > 0:
+            sockets = [sock for sock in sockets if sock._client_info["id"] in filter_ids]
+        res = [item.client_info_as_bytes for item in sockets]
+        return b"\n".join(res)
 
     @command(name="HELLO", fixed=(), repeat=(bytes,))
     def hello(self, *args: bytes) -> List[bytes]: